[fix] 修复 email 循环导入的问题
此修复解决了 email 模块中的循环导入问题，导入的模块从 "sign" 更改为 "_sign" 。循环导入可能导致代码执行错误或意外行为。通过此更改，我们避免了潜在的错误，并提高了代码的可靠性和一致性。
