+++ src/plugins/Core/plugins/email.py
@@ -1,8 +1,6 @@
 from ._utils import finish
 from .etm import items, data, bag
 from .su import su
-from .userinfo import panel
-from .sign import sign
 import os
 from nonebot import on_command, get_bot, get_bots
 from nonebot.matcher import Matcher
@@ -173,33 +171,6 @@ async def su_mail(event: MessageEvent, message: Message = CommandArg()) -> None:
                 await su.finish("完成！")
     except BaseException:
         await _error.report(traceback.format_exc(), su)
-
-
-@panel.handle()
-@sign.handle()
-async def unread_email_reminder(matcher: Matcher, event: MessageEvent):
-    try:
-        reminded_data = json.load(open("data/email.reminded.json", encoding="utf-8"))
-        if data.emails.get(event.get_user_id()):
-            email_count = 0
-            for email in data.emails[event.get_user_id()]:
-                if email not in (reminded_data.get(event.get_user_id()) or []):
-                    email_count += 1
-            if email_count != 0:
-                await matcher.send(
-                    _lang.text(
-                        "email.remind",
-                        [len(data.emails[event.get_user_id()])],
-                        event.get_user_id(),
-                    )
-                )
-                reminded_data[event.get_user_id()] = data.emails[event.get_user_id()]
-                json.dump(
-                    reminded_data,
-                    open("data/email.reminded.json", "w", encoding="utf-8"),
-                )
-    except BaseException:
-        await _error.report(traceback.format_exc(), matcher)
 
 
 @on_command("查看邮件", aliases={"view-emails", "ve", "ckyj"}).handle()
